applications/csv_agent/requirements.txt
langchain
langchain_core
langchain-google-genai
langchain_experimental
google-generativeai
python-dotenv
tabulate
streamlit